+++ src/constants.js
@@ -1,12 +1,16 @@
+const cssClasses = {
+    isHidden: 'is-hidden',
+    formError: 'form-error',
+    formErrors: 'form-errors',
+    hasError: 'has-error',
+    hasSuccess: 'has-success',
+    validationGroup: 'validation-group'
+};
+
 export default {
+    cssClasses,
     email: /^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$/,
     escapeChars: /[|\\{}()[\]^$+*?.]/g,
-    cssClasses: {
-        isHidden: 'is-hidden',
-        formError: 'form-error',
-        formErrors: 'form-errors',
-        hasError: 'has-error',
-        hasSuccess: 'has-success'
-    },
-    fieldValues: 'input, select, textarea, .validation-group'
+    fieldValues: `input, select, textarea, .${cssClasses.validationGroup}`,
+    validateOnOptions: ['blur', 'keyup']
 };