+++ CHANGELOG.md
@@ -4,6 +4,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.8.0
+------------------------------
+*January 24, 2018*
+
+### Added
+- Ability to validate on 'blur' or 'keyup'
+- Migrated 'dateinfuture' rule from other project
+- Explicitly restrict an implementation where you're able to both group messages and validate on blur/keyup
+
 v0.7.0
 ------------------------------
 *January 19, 2018*